Organize pages folder for better modularity
Organize the `src/pages` folder into subfolders for better modularity and maintainability.

* **Auth Pages**: Move `AuthCallback.jsx`, `EmailConfirmation.jsx`, `Login.jsx`, and `Registration.jsx` to `src/pages/auth`.
* **Website Pages**: Move `About.jsx`, `Contact.jsx`, `Home.jsx`, and `Pricing.jsx` to `src/pages/website`.
* **Console Pages**: Move `Integrations.jsx`, `Library.jsx`, `Workflow.jsx`, and `UserProfile.jsx` to `src/pages/console`.
* **Update Routes**: Modify import paths in `src/routes/index.jsx` to reflect the new file locations.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/lion-agi/lion-studio?shareId=XXXX-XXXX-XXXX-XXXX).
